requirements.txt
flask
flask-cors
tensorflow-cpu
pandas
numpy
scikit-learn
requests
gunicorn